Move yolov3-specific tasks into a yolov3 package
