+++ core/dsl/src/main/kotlin/edu/wpi/axon/dsl/task/yolov3/ConstructYoloV3ImageInput.kt
@@ -1,7 +1,8 @@
-package edu.wpi.axon.dsl.task
+package edu.wpi.axon.dsl.task.yolov3
 
 import edu.wpi.axon.dsl.Code
 import edu.wpi.axon.dsl.Import
+import edu.wpi.axon.dsl.task.Task
 import edu.wpi.axon.dsl.variable.Variable
 import edu.wpi.axon.util.singleAssign
